finished turnings longs and doubles into decimals, written IsInRange and used it where possible

+++ SaluteStocksAPI/Models/FundamentalData/IpoCalendar.cs
@@ -5,8 +5,8 @@ public class IpoCalendar
     [JsonProperty("symbol")] public string Symbol { get; set; }
     [JsonProperty("name")] public string Name { get; set; }
     [JsonProperty("ipoDate")] public DateTime? IpoDate { get; set; }
-    [JsonProperty("priceRangeLow")] public double? PriceRangeLow { get; set; }
-    [JsonProperty("priceRangeHigh")] public double? PriceRangeHigh { get; set; }
+    [JsonProperty("priceRangeLow")] public decimal? PriceRangeLow { get; set; }
+    [JsonProperty("priceRangeHigh")] public decimal? PriceRangeHigh { get; set; }
     [JsonProperty("currency")] public string Currency { get; set; }
     [JsonProperty("exchange")] public string Exchange { get; set; }
 }
